Media: Refactor the MediaUploadButton to be agnostic to its rendered UI
