+++ blocks/index.js
@@ -24,7 +24,8 @@ export { default as ColorPalette } from './color-palette';
 export { default as Editable } from './editable';
 export { default as EditableProvider } from './editable/provider';
 export { default as InspectorControls } from './inspector-controls';
-export { default as MediaUploadButton } from './media-upload-button';
+export { default as MediaUpload } from './media-upload';
+export { default as MediaUploadButton } from './media-upload/button';
 export { default as TermTreeSelect } from './term-tree-select';
 export { default as UrlInput } from './url-input';
 export { default as UrlInputButton } from './url-input/button';
